Issues/72 (#73)
* Added qr detected layout

* Added prompt screen layout

* Fixed prompt screen

* Added Edit QR Screen (Viewpager2)

* Fixed Edit QR Screen

* Added binding to Camera Class

* Added navigation from main activity to camera

* Refined main activity

* Created QR Detected Fragment java + Initialized navigation for camera activity

* Refined Camera code + Refined barcode scanning code

* Added navigation from camera to qr detected

* Refined Camera code

* Added javadoc

* Added bundle transfer rawValue from camera activity to QR detected activity

* Added hash name function + Added hash score function

* Set display for qr name and qr score

* Separated hash functions

* Refined hash functions + refined layouts

* Added Javadoc

* Fixed hash functions (bugs)

* Fixed layouts

* Added navigations from detected to prompt screen

* Refined camera codes (navigation + analyzer)

* Added photo-camera navigations

* Added navigation to prompt location screen

+++ code/app/src/main/java/com/example/qrquest/Info.java
@@ -45,7 +45,7 @@ public class Info extends QRCode {
      * @param score: the score of the QR Code
      * @param username: the username of the player
      */
-    public Info(String comment, String hashedQRCode, double latitude, double longitude, String qrCode, String region, double score, Date timeStamp, String username){
+    public Info(String comment, String hashedQRCode, double latitude, double longitude, String qrCode, String region, int score, Date timeStamp, String username){
         super(hashedQRCode, qrCode, score, latitude, longitude);
         setComment(comment);
         setRegion(region);
@@ -62,7 +62,7 @@ public class Info extends QRCode {
      * @param score: the score of the QR Code
      * @param username: the username of the player
      */
-    public Info(String comment, double latitude, double longitude, String qrCode, String region, double score, Date timeStamp, String username){
+    public Info(String comment, double latitude, double longitude, String qrCode, String region, int score, Date timeStamp, String username){
         super(qrCode, score, latitude, longitude);
         setRegion(region);
         setTimeStamp(timeStamp);
