+++ code/app/src/main/java/com/example/qrquest/QRCode.java
@@ -1,5 +1,6 @@
 package com.example.qrquest;
 
+
 /**
  * This class defines QR Codes
  */
@@ -11,14 +12,14 @@ public class QRCode {
     private String hashedQRCode;
 
     /**
-     * This member represents the unhashed name of the QR Code
+     * This member represents the un-hashed name of the QR Code
      */
     private String qrCode;
 
     /**
      * This member represents the value of the longitude of the QR Code
      */
-    private double score;
+    private int score;
 
     /**
      * This member represents the value of the latitude of the QR Code
@@ -38,12 +39,12 @@ public class QRCode {
     /**
      * This method defines an existing QR Code in the database
      * @param hashedQRCode: the hashed name of the QR Code
-     * @param qrCode: the unhashed name of the QR Code
+     * @param qrCode: the un-hashed name of the QR Code
      * @param score: the score of the QR Code
      * @param latitude: the latitude of the QR Code
      * @param longitude: the longitude of the QR Code
      */
-    public QRCode(String hashedQRCode, String qrCode, double score, double latitude, double longitude){
+    public QRCode(String hashedQRCode, String qrCode, int score, double latitude, double longitude){
         setHashedQRCode(hashedQRCode);
         setQrCode(qrCode);
         setScore(score);
@@ -53,12 +54,12 @@ public class QRCode {
 
     /**
      * This method defines a new QR Code
-     * @param qrCode: the unhashed name of the QR Code
+     * @param qrCode: the un-hashed name of the QR Code
      * @param score: the score of the QR Code
      * @param latitude: the latitude of the QR Code
      * @param longitude: the longitude of the QR Code
      */
-    public QRCode(String qrCode, double score, double latitude, double longitude){
+    public QRCode(String qrCode, int score, double latitude, double longitude){
         hashQRCode();
         setQrCode(qrCode);
         setScore(score);
@@ -67,16 +68,16 @@ public class QRCode {
     }
 
     /**
-     * This method retrieves the unhashed name of the QR Code
-     * @return the unhashed name of the QR Code
+     * This method retrieves the un-hashed name of the QR Code
+     * @return the un-hashed name of the QR Code
      */
     public String getQrCode() {
         return this.qrCode;
     }
 
     /**
-     * This method set the unhashed name for the QR Code and performs QR Code hashing
-     * @param qrCode: the unhashed name of the QR Code
+     * This method set the un-hashed name for the QR Code and performs QR Code hashing
+     * @param qrCode: the un-hashed name of the QR Code
      */
     public void setQrCode(String qrCode) {
         this.qrCode = qrCode;
@@ -117,7 +118,7 @@ public class QRCode {
      * This method sets the score for the QR Code
      * @param score: the score of the QR Code
      */
-    public void setScore(double score) {
+    public void setScore(int score) {
         this.score = score;
     }
